requirements.txt
numpy>=1.26.4
torch>=2.2.2
torchvision>=0.17.2
scipy>=1.14.0
pytz>=2024.1
thop>=0.1.1-2209072238
kornia>=0.7.3
timm>=1.0.7
einops>=0.8.0
tensorboardX>=2.2
PyYAML>=6.0.1
opencv-python>=4.10.0.84
matplotlib>=3.9.1
tqdm>=4.66.4
pillow>=10.4.0